Move finalize command to block protocol server
