+++ hadoop-hdds/framework/src/main/java/org/apache/hadoop/hdds/scm/client/ScmClient.java
@@ -473,8 +473,6 @@ public interface ScmClient extends Closeable {
       String upgradeClientID, boolean force, boolean readonly)
       throws IOException;
 
-  HddsProtos.UpgradeStatus queryUpgradeStatus() throws IOException;
-
   DecommissionScmResponseProto decommissionScm(
       String scmId) throws IOException;
