+++ hadoop-hdds/framework/src/main/java/org/apache/hadoop/hdds/scm/protocol/StorageContainerLocationProtocol.java
@@ -518,23 +518,10 @@ public interface StorageContainerLocationProtocol extends Closeable {
   StatusAndMessages finalizeScmUpgrade(String upgradeClientID)
       throws IOException;
 
-  void finalizeUpgrade() throws IOException;
-
-  /**
-   * Same as {@link #finalizeUpgrade()}, but SCM skips the peer SCM and datanode software version
-   * checks before finalizing. Use this to finalize when a peer or datanode is intentionally down or
-   * on a different version.
-   *
-   * @throws IOException If any error occurs.
-   */
-  void forceFinalizeUpgrade() throws IOException;
-
   @Deprecated
   StatusAndMessages queryUpgradeFinalizationProgress(
       String upgradeClientID, boolean force, boolean readonly)
       throws IOException;
-
-  HddsProtos.UpgradeStatus queryUpgradeStatus() throws IOException;
 
   /**
    * Returns the {@link HDDSVersion#SOFTWARE_VERSION} of the responding SCM binary. Intended for the
