+++ hadoop-hdds/interface-server/src/main/proto/ScmServerProtocol.proto
@@ -39,6 +39,8 @@ enum Type {
   SortDatanodes      = 14;
   AddScm             = 15;
   GetClusterTree     = 16;
+  FinalizeUpgrade    = 17;
+  QueryUpgradeStatus = 18;
 }
 
 message SCMBlockLocationRequest {
@@ -57,6 +59,8 @@ message SCMBlockLocationRequest {
   optional SortDatanodesRequestProto          sortDatanodesRequest      = 14;
   optional hadoop.hdds.AddScmRequestProto     addScmRequestProto       = 15;
   optional GetClusterTreeRequestProto         getClusterTreeRequest     = 16;
+  optional FinalizeUpgradeRequestProto        finalizeUpgradeRequest    = 17;
+  optional QueryUpgradeStatusRequestProto     queryUpgradeStatusRequest = 18;
 }
 
 message SCMBlockLocationResponse {
@@ -82,6 +86,8 @@ message SCMBlockLocationResponse {
   optional SortDatanodesResponseProto          sortDatanodesResponse      = 14;
   optional hadoop.hdds.AddScmResponseProto     addScmResponse        = 15;
   optional GetClusterTreeResponseProto         getClusterTreeResponse     = 16;
+  optional FinalizeUpgradeResponseProto        finalizeUpgradeResponse    = 17;
+  optional QueryUpgradeStatusResponseProto     queryUpgradeStatusResponse = 18;
 }
 
 /**
@@ -244,6 +250,20 @@ message GetClusterTreeRequestProto {
 
 message GetClusterTreeResponseProto {
   required InnerNode clusterTree = 1;
+}
+
+message FinalizeUpgradeRequestProto {
+  optional bool force = 1 [default = false];
+}
+
+message FinalizeUpgradeResponseProto {
+}
+
+message QueryUpgradeStatusRequestProto {
+}
+
+message QueryUpgradeStatusResponseProto {
+  required hadoop.hdds.UpgradeStatus status = 1;
 }
 
 /**
