+++ hadoop-hdds/server-scm/src/main/java/org/apache/hadoop/hdds/scm/server/SCMClientProtocolServer.java
@@ -93,12 +93,10 @@ import org.apache.hadoop.hdds.scm.container.reconciliation.ReconciliationEligibi
 import org.apache.hadoop.hdds.scm.events.SCMEvents;
 import org.apache.hadoop.hdds.scm.exceptions.SCMException;
 import org.apache.hadoop.hdds.scm.exceptions.SCMException.ResultCodes;
-import org.apache.hadoop.hdds.scm.ha.SCMNodeDetails;
 import org.apache.hadoop.hdds.scm.ha.SCMRatisServer;
 import org.apache.hadoop.hdds.scm.ha.SCMRatisServerImpl;
 import org.apache.hadoop.hdds.scm.node.DatanodeInfo;
 import org.apache.hadoop.hdds.scm.node.DatanodeUsageInfo;
-import org.apache.hadoop.hdds.scm.node.NodeManager;
 import org.apache.hadoop.hdds.scm.node.NodeStatus;
 import org.apache.hadoop.hdds.scm.node.states.NodeNotFoundException;
 import org.apache.hadoop.hdds.scm.pipeline.Pipeline;
@@ -107,10 +105,8 @@ import org.apache.hadoop.hdds.scm.pipeline.PipelineManager;
 import org.apache.hadoop.hdds.scm.pipeline.PipelineNotFoundException;
 import org.apache.hadoop.hdds.scm.protocol.StorageContainerLocationProtocol;
 import org.apache.hadoop.hdds.scm.protocol.StorageContainerLocationProtocolServerSideTranslatorPB;
-import org.apache.hadoop.hdds.scm.protocolPB.StorageContainerLocationProtocolClientSideTranslatorPB.ScmNodeTarget;
 import org.apache.hadoop.hdds.scm.protocolPB.StorageContainerLocationProtocolPB;
 import org.apache.hadoop.hdds.security.SecurityConfig;
-import org.apache.hadoop.hdds.utils.HAUtils;
 import org.apache.hadoop.hdds.utils.HddsServerUtil;
 import org.apache.hadoop.hdds.utils.ProtocolMessageMetrics;
 import org.apache.hadoop.io.IOUtils;
@@ -1170,99 +1166,8 @@ public class SCMClientProtocolServer implements
   }
 
   @Override
-  public void finalizeUpgrade() throws IOException {
-    finalizeUpgrade(false);
-  }
-
-  @Override
-  public void forceFinalizeUpgrade() throws IOException {
-    finalizeUpgrade(true);
-  }
-
-  private void finalizeUpgrade(boolean force) throws IOException {
-    final Map<String, String> auditMap = Collections.singletonMap("force", String.valueOf(force));
-    try {
-      getScm().checkAdminAccess(getRemoteUser(), false);
-      if (force) {
-        LOG.warn("Forcing upgrade finalization by skipping SCM peer and datanode software version checks");
-      } else {
-        validatePeerScmVersionsBeforeFinalize();
-        validateDatanodeVersionsBeforeFinalize();
-      }
-      scm.getFinalizationManager().finalizeUpgrade();
-      AUDIT.logWriteSuccess(buildAuditMessageForSuccess(SCMAction.FINALIZE_SCM_UPGRADE, auditMap));
-    } catch (Exception ex) {
-      AUDIT.logWriteFailure(buildAuditMessageForFailure(SCMAction.FINALIZE_SCM_UPGRADE, auditMap, ex));
-      throw ex;
-    }
-  }
-
-  @Override
   public HDDSVersion getPeerUpgradeStatus() throws IOException {
     return HDDSVersion.SOFTWARE_VERSION;
-  }
-
-  /**
-   * Verifies that every peer SCM in the Ratis group runs the same software version as this leader
-   * before finalization begins. Rejects the finalize command if any peer reports a differing
-   * version or cannot be reached. The resulting exception propagates back to the OM (which triggered
-   * finalization) and on to the client, leaving nothing finalized.
-   */
-  private void validatePeerScmVersionsBeforeFinalize() throws SCMException {
-    List<SCMNodeDetails> peerNodes = scm.getSCMHANodeDetails().getPeerNodeDetails();
-    if (peerNodes.isEmpty()) {
-      return;
-    }
-    HDDSVersion leaderVersion = HDDSVersion.SOFTWARE_VERSION;
-    OzoneConfiguration conf = scm.getConfiguration();
-    List<String> failedPeers = new ArrayList<>();
-    for (SCMNodeDetails peer : peerNodes) {
-      String peerId = peer.getNodeId();
-      ScmNodeTarget target = new ScmNodeTarget();
-      target.setNodeId(peerId);
-      StorageContainerLocationProtocol peerClient = null;
-      try {
-        // Contact the peer SCM as this SCM's service (Kerberos keytab) identity, not the remote client's identity.
-        // This runs inside the finalize RPC handler's doAs context, whose UGI has no credentials to open a fresh
-        // outbound RPC to the peer SCM.
-        peerClient = HAUtils.getScmContainerClientForNode(conf, target, UserGroupInformation.getLoginUser());
-        HDDSVersion peerVersion = peerClient.getPeerUpgradeStatus();
-        if (!peerVersion.equals(leaderVersion)) {
-          LOG.warn("SCM peer {} is running software version {} but leader is running version {}. "
-              + "Rejecting finalize command.", peerId, peerVersion, leaderVersion);
-          failedPeers.add(peerId + " (version: " + peerVersion + ")");
-        }
-      } catch (IOException e) {
-        LOG.warn("Failed to contact SCM peer {} to check software version before finalize.", peerId, e);
-        failedPeers.add(peerId + " (unreachable: " + e.getMessage() + ")");
-      } finally {
-        IOUtils.cleanupWithLogger(LOG, peerClient);
-      }
-    }
-    if (!failedPeers.isEmpty()) {
-      throw new SCMException("Finalize rejected: the following SCM peers did not confirm matching software "
-          + "version (expected version=" + leaderVersion + "): " + String.join(", ", failedPeers),
-          ResultCodes.UNSUPPORTED_OPERATION);
-    }
-  }
-
-  /**
-   * Verifies that every healthy datanode runs the same software version as this SCM before
-   * finalization begins. Datanodes finalize only after SCM instructs them to, so this does not
-   * require them to be finalized; it requires their binaries to match SCM's software version.
-   * Rejects the finalize command if any healthy datanode reports a differing (or unknown) software
-   * version. The resulting exception propagates back to the OM (which triggered finalization) and on
-   * to the client, leaving nothing finalized.
-   */
-  private void validateDatanodeVersionsBeforeFinalize() throws SCMException {
-    NodeManager.DatanodeFinalizationCounts counts =
-        scm.getScmNodeManager().getDatanodeFinalizationCounts();
-    if (!counts.allSoftwareVersionsMatchScmVersion()) {
-      LOG.warn("Rejecting finalize command: not all {} healthy datanodes are running SCM's software "
-          + "version {}.", counts.getTotalHealthyDatanodes(), HDDSVersion.SOFTWARE_VERSION);
-      throw new SCMException("Finalize rejected: not all healthy datanodes are running the SCM software version "
-          + HDDSVersion.SOFTWARE_VERSION, ResultCodes.UNSUPPORTED_OPERATION);
-    }
   }
 
   @Override
@@ -1293,56 +1198,6 @@ public class SCMClientProtocolServer implements
     } catch (IOException ex) {
       AUDIT.logReadFailure(buildAuditMessageForFailure(
           SCMAction.QUERY_UPGRADE_FINALIZATION_PROGRESS, auditMap, ex));
-      throw ex;
-    }
-  }
-
-  @Override
-  public HddsProtos.UpgradeStatus queryUpgradeStatus() throws IOException {
-    try {
-      getScm().checkAdminAccess(getRemoteUser(), true);
-
-      if (scm.getScmContext().isInSafeMode()) {
-        throw new SCMException("Cannot query upgrade status while SCM is in safe mode. Wait until SCM exits "
-            + "safe mode and try again.", ResultCodes.SAFE_MODE_EXCEPTION);
-      }
-
-      // Set SCM finalization status to return to the client.
-      // Since SCM finalization goes through Ratis, it moves from unfinalized to finalized immediately with no
-      // in-progress state.
-      boolean scmFinalized = !scm.getVersionManager().needsFinalization();
-      HddsProtos.FinalizationStatus scmFinalizationStatus =
-          scmFinalized ? HddsProtos.FinalizationStatus.FINALIZED : HddsProtos.FinalizationStatus.UNFINALIZED;
-
-      // Set overall HDDS finalization status (SCM and Datanodes) to return to the client.
-      NodeManager.DatanodeFinalizationCounts datanodeFinalizationCounts =
-          scm.getScmNodeManager().getDatanodeFinalizationCounts();
-      int finalizedDatanodes = datanodeFinalizationCounts.getNumFinalizedDatanodes();
-      int healthyDatanodes = datanodeFinalizationCounts.getTotalHealthyDatanodes();
-      HddsProtos.FinalizationStatus hddsFinalizationStatus;
-      if (!scmFinalized) {
-        // SCM must finish finalizing before Datanodes can start finalizing.
-        hddsFinalizationStatus = HddsProtos.FinalizationStatus.UNFINALIZED;
-      } else if (datanodeFinalizationCounts.allNodesFinalized()) {
-        hddsFinalizationStatus = HddsProtos.FinalizationStatus.FINALIZED;
-      } else {
-        hddsFinalizationStatus = HddsProtos.FinalizationStatus.IN_PROGRESS;
-      }
-
-      HddsProtos.UpgradeStatus result = HddsProtos.UpgradeStatus.newBuilder()
-          .setScmFinalizationStatus(scmFinalizationStatus)
-          .setNumDatanodesFinalized(finalizedDatanodes)
-          .setNumDatanodesTotal(healthyDatanodes)
-          .setHddsFinalizationStatus(hddsFinalizationStatus)
-          .setScmApparentVersion(scm.getVersionManager().getApparentVersion().serialize())
-          .setMinDatanodeApparentVersion(datanodeFinalizationCounts.getMinApparentVersion())
-          .setMaxDatanodeApparentVersion(datanodeFinalizationCounts.getMaxApparentVersion())
-          .build();
-
-      AUDIT.logReadSuccess(buildAuditMessageForSuccess(SCMAction.QUERY_UPGRADE_STATUS, null));
-      return result;
-    } catch (IOException ex) {
-      AUDIT.logReadFailure(buildAuditMessageForFailure(SCMAction.QUERY_UPGRADE_STATUS, null, ex));
       throw ex;
     }
   }
