+++ hadoop-ozone/cli-admin/src/main/java/org/apache/hadoop/hdds/scm/cli/ContainerOperationClient.java
@@ -615,11 +615,6 @@ public class ContainerOperationClient implements ScmClient {
   }
 
   @Override
-  public HddsProtos.UpgradeStatus queryUpgradeStatus() throws IOException {
-    return storageContainerLocationClient.queryUpgradeStatus();
-  }
-
-  @Override
   public DecommissionScmResponseProto decommissionScm(
       String scmId)
       throws IOException {
